+++ lib/app/data/models/money_objects/loan_payments/loan_payments.dart
@@ -4,6 +4,7 @@ import 'package:money/app/data/models/money_objects/accounts/account.dart';
 import 'package:money/app/data/models/money_objects/accounts/account_types_enum.dart';
 import 'package:money/app/data/models/money_objects/loan_payments/loan_payment.dart';
 import 'package:money/app/data/models/money_objects/money_objects.dart';
+import 'package:money/app/data/models/money_objects/transactions/transaction.dart';
 import 'package:money/app/data/storage/data/data.dart';
 
 // Export
@@ -77,7 +78,7 @@ List<LoanPayment> getAccountLoanPayments(Account account) {
       .toList();
 
   // include the bank transactions matching th Account Categories for Principal and Interest
-  var listOfTransactions = Data().transactions.getListFlattenSplits(
+  final List<Transaction> listOfTransactions = Data().transactions.getListFlattenSplits(
         whereClause: (t) => t.isMatchingAnyOfTheseCategoris(categoriesToMatch),
       );