+++ lib/app/data/models/money_objects/transactions/transaction_types.dart
@@ -9,7 +9,7 @@ enum TransactionStatus {
 String transactionStatusToLetter(final TransactionStatus status) {
   switch (status) {
     case TransactionStatus.none:
-      return 'N';
+      return '_';
     case TransactionStatus.electronic:
       return 'E';
     case TransactionStatus.cleared: